requirements.txt
django
django-jazzmin
django-filter
django-jsonfield
django-nested-admin
djangorestframework
djangorestframework-simplejwt
drf-yasg